% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_calling.R
\name{arg_calling}
\alias{arg_calling}
\title{Pull out an argument calling a specified function}
\usage{
arg_calling(ex, n = 1L, fail = "", qfuns)
}
\arguments{
\item{ex}{An call to be checked}

\item{n}{Look for the nth passing argument}

\item{fail}{character string message on failure}

\item{qfuns}{A quoted function name or vector of names}
}
\description{
Find an argument
}
\examples{
ex <- for_checkr(quote(sin(53 * pi / 180)))
arg_calling(ex, qfuns = quote(sin))

}
