% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tests.R
\name{assigns}
\alias{assigns}
\title{Basic tests for single expressions}
\usage{
assigns(ex, target = TRUE)
}
\arguments{
\item{ex}{a single expression, e.g. a <- sin(x)}

\item{target}{the anticipated match}
}
\value{
checkr result if no match,
otherwise the rhs of the expression
}
\description{
\code{assigns} checks whether the expression is an assignment.
If target is a character string, additionally checks whether the
target of assignment matches the character string.

These handle just one expression, not a bracketed set.
}
\examples{
ex1 <- quote(x <- sin(a + b))
ex2 <- quote(sin(a + b))
assigns(ex1)
assigns(ex2)
assigns(ex1, target = FALSE)
assigns(ex2, target = FALSE) # shouldn't be assignment
assigns(ex1, target = x)
assigns(ex1, target = y)
}
