% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_learnr.R
\name{check_for_learnr}
\alias{check_for_learnr}
\title{Interface for checking learnr exercises.}
\usage{
check_for_learnr(label = NULL, user_code = NULL, solution_code = NULL,
  check_code = NULL, envir_result = NULL, evaluate_result = NULL, ...,
  debug = FALSE)
}
\description{
When using learnr, the check statements should \emph{NOT} be wrapped with \code{check().}
Instead, put the \code{if_matches()} functions directly in the check-chunk. The \code{ex} argument
to \code{if_matches()} should be literally \code{USER_CODE}.
}
\examples{
# as it would be called from the learnr system ...
check_for_learnr(label = "first", user_code = "sin(pi)",
  check_code = 'if_matches(USER_CODE, .(fn)(..(a)), passif(fn \%same_as\% as.name("sin") && a == pi, message="Right-o!"))')


}
