% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%or\%}
\alias{\%or\%}
\alias{\%and\%}
\alias{The}
\alias{outcome}
\alias{of}
\alias{a}
\alias{checkr}
\alias{test}
\alias{is}
\alias{an}
\alias{object}
\alias{class}
\alias{`checkr_result`}
\alias{that}
\alias{indicates}
\alias{whether}
\alias{the}
\alias{was}
\alias{passed}
\alias{or}
\alias{failed}
\alias{sufficient}
\alias{to}
\alias{move}
\alias{on}
\alias{additional}
\alias{tests.}
\title{Utilities for checkr results}
\usage{
res1 \%or\% res2
}
\arguments{
\item{res1}{the first checkr result}

\item{res2}{the second checkr result}
}
\description{
Utilities for checkr results
}
\examples{
# normally these results are made by the checking functions
pa <- checkr2:::new_checkr_result("pass", message = "Good!")
pb <- checkr2:::new_checkr_result("pass", message = "Great!")
fa <- checkr2:::new_checkr_result("fail", message = "too bad")
fb <- checkr2:::new_checkr_result("fail", message = "sorry")
oka <- checkr2:::new_checkr_result("ok", message = "note a")
okb <- checkr2:::new_checkr_result("ok", message = "note b")
pa \%or\% pb
pa \%and\% pb
fa \%or\% fb
fa \%and\% fb
oka \%or\% okb
oka \%and\% okb
oka \%and\% fa
oka \%or\% fa
pa \%or\% fa
pa \%and\% fa
pa \%or\% oka
pa \%and\% oka

}
