% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{curly_to_tidy}
\alias{curly_to_tidy}
\title{Translate a sequence of commands into a sequence of quosures}
\usage{
curly_to_tidy(exprs)
}
\arguments{
\item{exs}{an expression or expressions in a curly brace}
}
\value{
a quosure for each expression in exs, with an associated
environment that reflects the state when that expression was evaluated.
}
\description{
Translate a sequence of commands into a sequence of quosures
}
\examples{
code <- curly_to_tidy(quote({x <- 2; y <- x^2; z <- x + y}))
at(code, Z == "z")
at(code, V == 4, Z == "y")
if_matches(at(code, F == `+`), `+`(.(one), .), req(one != quote(x), "fated message for {{one}}"))
}
