% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_if_no_match.R
\name{fail_if_no_match}
\alias{fail_if_no_match}
\title{Check for missing statements}
\usage{
fail_if_no_match(ex, key, message = "Pattern not found.")
}
\arguments{
\item{ex}{A set of expressions: the submission}

\item{key}{A pattern to match to the expressions}

\item{message}{The message to return}
}
\description{
Match the pattern against all statements in the expressions. If none of the
statements matches the pattern, return the message as a fail.
}
\examples{
# check for assignment to a given name
USER_CODE <- quote({z <- 1; foo <- z})
fail_if_no_match(USER_CODE, x <- ., "You didn't give any value to `x`.")
}
