% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tests.R
\name{generic_arg}
\alias{generic_arg}
\title{Functions that look at the arguments of a
matching call, to make sure it meets certain criteria.
In the tests, the expression found can be referred to as QQ
and its value as VV.}
\usage{
generic_arg(type, type_description, ex, ...)
}
\arguments{
\item{ex}{the expression to check}

\item{...}{specific tests}

\item{nm}{the name of an argument. Can be bare or back-ticked.}
}
\description{
Functions that look at the arguments of a
matching call, to make sure it meets certain criteria.
In the tests, the expression found can be referred to as QQ
and its value as VV.
}
\examples{
ex <- quote(lm(mpg ~ hp, data = mtcars))
formula_arg(ex)
formula_arg(ex, passif(is_formula(QQ), "Found it!"))
formula_arg(ex, failif(f_rhs(QQ) != quote(mpg), "You should have 'mpg' on the right side"))
}
