% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{is_chain}
\alias{is_chain}
\title{#' @export
chain_element <- function(tidy_code, n = 1L) {
stopifnot(inherits(tidy_code, "checkr_result"))
stopifnot(is_chain(tidy_code$code[[1]]))

whole_chain <- tidy_code$code[[1]]
return(whole_chain)
}

[[1]: R:[1
[[1]: R:[1}
\usage{
is_chain(ex)
}
\description{
#' @export
chain_element <- function(tidy_code, n = 1L) {
stopifnot(inherits(tidy_code, "checkr_result"))
stopifnot(is_chain(tidy_code$code[[1]]))

whole_chain <- tidy_code$code[[1]]
return(whole_chain)
}

[[1]: R:[1
[[1]: R:[1
}
