% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{line_after}
\alias{line_after}
\title{Grab the lines after a specified line (which is included)}
\usage{
line_after(tidy_code, ..., fail = "", type = NULL)
}
\description{
Grab the lines after a specified line (which is included)
}
