% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chaining.R
\name{line_chaining}
\alias{line_chaining}
\title{Find a line that is a magrittr chain}
\usage{
line_chaining(ex, n = 1L)
}
\arguments{
\item{ex}{a checkr_result object such as produced by \code{for_checkr()}}

\item{n}{An integer indicating which line that is a chain is desired: first, second, ...}
}
\description{
Code submissions may consist of one or more lines. Each of those lines may be
written as a magrittr chain. \code{line_chaining()} will extract a single line that is a chain.
}
\details{
A chain is itself equivalent to multiple lines. If you are interested in
exploring within a chain, you can expand the chain into individual lines. See \code{\link[checkr2:expand_chains]{checkr2::expand_chains()}}
}
\seealso{
\code{\link[checkr2:expand_chains]{checkr2::expand_chains()}}, \code{\link[checkr2:expand_all_chains]{checkr2::expand_all_chains()}}
}
