% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{lines_after}
\alias{lines_after}
\title{Grab the lines after a specified line (which is included)}
\usage{
lines_after(tidy_code, ..., message = "")
}
\description{
Grab the lines after a specified line (which is included)
}
