% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misconception.R
\name{misconception}
\alias{misconception}
\title{test for a misconceived statement}
\usage{
misconception(tidy_code, pattern, message)
}
\arguments{
\item{tidy_code}{A checkr result to use as input}

\item{pattern}{A code pattern as found by any of the functions returning a checkr_result object.}

\item{message}{The message to give if the \code{pattern} is found.}
}
\description{
Sometimes you expect a specific code pattern which might reflect a common misconception.
\code{misconception()} let's you mark a pattern as such, causing the test to fail if the pattern
is found.
}
