% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moustache.R
\name{moustache}
\alias{moustache}
\title{evaluate expressions in a string using a particular environment}
\usage{
moustache(string, bindings = env_parent())
}
\description{
evaluate expressions in a string using a particular environment
}
\examples{
moustache("hello")
moustache("Three plus five is {{3+5}}.")
}
