% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_EX.R
\name{n_args}
\alias{n_args}
\title{Tests on expressions}
\usage{
n_args(EX)
}
\description{
These functions can be applied to EX statements as produced by \code{chk()}.
They calculate something about the expression.
}
