% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_check.R
\name{parse_check}
\alias{parse_check}
\title{Look for errors at parse time}
\usage{
parse_check(user_code)
}
\arguments{
\item{user_code}{a character string containing the code to parse}
}
\description{
Look for errors at parse time
}
\examples{
code <- "lm(mpg ~ hp, \\ndata = mtcars); \\nplot((1:10)\\n x <- f[b]\\n y <- x^2\\n\\n z = cos(yy * xx^2)"
parse_check(code)
code2 <- "__blank__ <- 2"
cat(parse_check(code2)$message)
}
