% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_anyways.R
\name{pass_anyways}
\alias{pass_anyways}
\title{Produce a passing test result with a message.}
\usage{
pass_anyways(message = "Good!")
}
\arguments{
\item{message}{a character string giving the message for success.}
}
\description{
This is used to guarantee a pass result if no other checking function returns a
pass or fail.
}
