% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{generic_test}
\alias{generic_test}
\alias{passif}
\alias{failif}
\alias{noteif}
\title{Collate a test expression with a message and a action.}
\usage{
passif(test, message)
}
\arguments{
\item{test}{an expression written in terms of values found
in the pattern-matching bindings}

\item{message}{a character string containing the message to return.}
}
\description{
These functions are simply a way to associate a message with a test.
The test itself will be evaluated in \code{run_tests()}, not in these functions.
}
\details{
The \code{message} can include components calculated from the bindings.
Enclose these in moustaches, e.g. "The \code{{{fn}}} function is not appropriate for adding."
Within a test, the operators \code{==} and \code{!=} in a test have been augmented to deal
with language objects such as names. They are translated to be equivalent to "\%same_as\%".
}
