% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_check.R
\name{pre_check}
\alias{pre_check}
\title{Pre-check code}
\usage{
pre_check(user_code, soln_code = "hello")
}
\description{
Pre-check code
}
\examples{
code <- "lm(mpg ~ hp, data <- mtcars); plot(1:10); x <- 1\\n y <- x^2\\n\\n z = cos(yy * xx^2)"
pre_check(code)


}
