% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_ok.R
\name{resolve_ok}
\alias{resolve_ok}
\title{Translate an 'ok' message into a pass or fail}
\usage{
resolve_ok(test_result, pass = TRUE, message = "Good!")
}
\arguments{
\item{test_result}{The outcome of some checkr test}

\item{pass}{If \code{TRUE}, the ok will be turned into a pass. If \code{FALSE}, a failure.}

\item{message}{Character string message to include with the transformed result.}
}
\description{
Allows you to force an OK message to be either a pass or fail.
}
\details{
Sometimes "ok" test results come with a message of its own. The message given as an argument
will be pre-pended to whatever message was with the "ok" test result.
}
\examples{
CODE <- for_checkr("sin(pi)")
L1 <- line_where(CODE, F != sin) # gives an OK, because `fail` was not set.
resolve_ok(L1)


}
