% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{line_at}
\alias{line_at}
\alias{lines_after}
\title{Identify an individual line in a sequence of commands}
\usage{
line_at(tidy_code, ..., fail = "", type = NULL)
}
\arguments{
\item{tidy_code}{expressions as made by for_checkr()}

\item{...}{tests specifying the kind of line we want}

\item{fail}{if a non-empty string, trigger a failure if no matching
line is found, with the string as the message.}

\item{type}{a test to check whether V is a certain type, e.g. \code{is.numeric}.}
}
\description{
The tests are written in terms of pronouns
\itemize{
\item F the function at the highest level (but with assignment removed)
\item Z the name being bound to the line's value by assignment. ("" if no assignment.)
\item V the value produced by the line.
\item EX the expression itself (but with assignment removed)
}
}
\details{
\code{fail} and \code{type} are optional. The \code{fail} argument merely sets the
message if no matching line is found. \code{type} can be used to ensure that the value
V is an appropriate kind for the tests specified in ...
}
\examples{
tidy_code <- for_checkr(quote({x <- 2; y <- x^3; z <- y + x}))
line_at(tidy_code, F \%same_as\% quote(`^`))

}
