% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_calling.R
\name{arg_calling}
\alias{arg_calling}
\title{Pull out an line, expression, or argument calling a specified function}
\usage{
arg_calling(ex, ..., n = 1L, message = "call to function")
}
\arguments{
\item{ex}{An call to be checked}

\item{...}{unquoted function names}

\item{n}{Look for the nth passing argument}

\item{fail}{character string message on failure}

\item{just_the_fun}{Flag for internal use.}
}
\description{
Find an argument
}
\examples{
ex <- for_checkr(quote(15 * sin(53 * pi / 180)))
ex2 <- for_checkr(quote(sin(3)))
line_calling(ex, qfuns = quote(sin))
# look everywhere, including the top-level function
call_to(ex, qfuns = quote(sin))
call_to(ex2, qfuns = quote(sin))
# look only in arguments: the top-level function doesn't count
arg_calling(ex, qfuns = quote(sin))
arg_calling(ex2, qfuns = quote(sin))

}
