% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Apply passif/failif/okif tests to an identified code object}
\usage{
check(tidy_code_line, ...)
}
\arguments{
\item{...}{passif/failif tests to be applied}

\item{ex}{the checkr_result object with a single line of code}
}
\value{
a checkr_result object reflecting the outcome of the tests
}
\description{
A generic testing function for applying passif/failif tests with V and EX bindings
to a checkr_result object. Unlike \code{line_where()}, other line functions, no patterns
need to be provided. Just the passif/failif tests.
}
