% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blanks.R
\name{check_blanks}
\alias{check_blanks}
\title{Checking fill-in-the-blanks problems}
\usage{
check_blanks(ex, pat, ...)
}
\arguments{
\item{ex}{the expression to check, e.g. USER_CODE}

\item{pat}{the anticipated correct form. Use elements such as \code{..one..} to
define a blank.}

\item{...}{tests to apply to expressions in \code{ex}. These are typically made
with \code{passif()}, \code{failif()}, and so on.}
}
\description{
Checking fill-in-the-blanks problems
}
\examples{
submission <- quote(res <- sqrt(a^2 + b^2))
submission2 <- quote(res <- sin(a^2 + b^2))
as_posted <- quote({res <- ..fn..(a^2 + b^2)})
check_blanks(submission2, !!as_posted,
   failif(fn == quote(sin) , "Wrong function"),
   passif(fn == quote(sqrt), "Right: the square root!"))
as_posted <- quote({res <- ..fn..(`+`(`^`(a, ..exp1..), `^`(b, ..exp2..)))})
}
