% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_bindings.R
\name{grab_bindings}
\alias{grab_bindings}
\title{Grab any bindings from a pattern match}
\usage{
grab_bindings(ex, keys, req_all_patterns = TRUE, fail_if_no_match = TRUE)
}
\arguments{
\item{ex}{the expression to be searched for the patterns}

\item{keys}{the set of patterns}
}
\description{
This is too similar to the code in \code{if_matches()}.
}
\examples{
ex <- quote(plot(mpg ~ wt, data = subset(mtcars, hp < 150)))
key <- quote(.(fn)(.(formula), data = .(the_data)))
grab_bindings(ex, !!key, fail_if_no_match = FALSE)
ex2 <- quote({x <- 1; y <- x^2})
keys <- quote({`<-`(x, ..(first)); `<-`(.(var), .(fn)(x, .(exponent)))})
grab_bindings(ex2, !!keys, fail_if_no_match = FALSE)
grab_bindings(ex2, !!keys, fail_if_no_match = FALSE, req_all_patterns = FALSE)

}
