% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_anyways.R
\name{pass_anyways}
\alias{pass_anyways}
\alias{fail_anyways}
\title{Produce a test result definitively.}
\usage{
pass_anyways(message = "Good!")
}
\arguments{
\item{message}{a character string giving the message for success.}
}
\description{
These are used to guarantee a pass or fail result if no other checking function returns a
pass or fail.
}
